refactor(cache): put the remote client behind a backend enum
`RemoteCacheClient` keeps its name, its constructor, and every method
signature, but is now a facade over a crate-private `Backend` enum whose
single variant holds the HTTP protocol client. Nothing outside this crate
changes: the agent, the upload queue, the session, and doctor all call the
same methods on the same type.

Backends are a closed set, so they dispatch through an enum rather than a
trait. An `async fn` in a trait is not object-safe, and boxing every call to
work around that would buy nothing when the alternative is a `match` the
compiler can see through.

The HTTP implementation moves to `remote_http.rs` unchanged -- endpoints,
capability negotiation, blob-pack framing, credentials, and the URL
validation that only a protocol server needs. Helpers both backends will
want stay in `lib.rs`: the bounded JSON reader, the entity-tag parser, and
the retry loop. Method documentation moves to the facade, which is the
public surface a reader lands on.

The one thing that is not a pure move: three doc comments had drifted onto
`PackMemberSource`, one of them a duplicate of the paragraph already above
`blob_pack_stream` and another belonging to `read_bounded_json`. They are
restored to what they describe.

`core_tests.rs` becomes the test module of `remote_http.rs`, since that is
where what it exercises now lives. Its only edits are the `use` lines and
the two helpers that reach for endpoint builders rather than the protocol.
